+++ examples/accounts/accounts.tsx
@@ -66,7 +66,6 @@ export const accounts: AccountDataProps[] = [
     name: 'Ståle Solbakken',
     uniqueId: '071095XXXXX',
     emailAlerts: false,
-    disabled: true,
     smsAlerts: true,
   },
   {
@@ -138,7 +137,6 @@ export const accounts: AccountDataProps[] = [
     name: 'Vålerenga fotball',
     uniqueId: '968218743',
     isDeleted: true,
-    disabled: true,
   },
   {
     id: 'vif-2',
@@ -147,7 +145,6 @@ export const accounts: AccountDataProps[] = [
     uniqueId: '972202223',
     parentId: 'vif',
     isDeleted: true,
-    disabled: true,
   },
   {
     id: 'sogndal',
@@ -169,6 +166,7 @@ export const accounts: AccountDataProps[] = [
     type: 'company',
     name: 'Stiftelsen Kanvas',
     uniqueId: '971272643',
+    disabled: true,
   },
   {
     id: 'abra-cadabra',